movie_chatbot/requirements.txt
# Core
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
jinja2==3.1.2

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0

# Database
sqlalchemy==2.0.19
pymongo==4.5.0
alembic==1.11.3

# Scheduler
apscheduler==3.10.1

# Data Processing
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2

# Authentication
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.0

# Development
black==23.3.0
isort==5.12.0
flake8==6.0.0
mypy==1.3.0